Review comments fixx. Code improvement.,
